Move UtcDateTime unit tests from Parse/ to misc/.
